Add support for profiling to `dns` binary
Adding profiling support to `dns` binary to justify future performance
improvements.
