+++ mtop/src/bin/mc.rs
@@ -1,14 +1,12 @@
 use clap::{Args, Parser, Subcommand, ValueHint};
 use mtop::bench::{Bencher, Percent, Summary};
 use mtop::check::{Checker, TimingBundle};
-use mtop::profile::Profiler;
+use mtop::profile;
 use mtop_client::{
     DiscoveryDefault, MemcachedClient, MemcachedPool, Meta, MtopError, PoolConfig, SelectorRendezvous, Server,
     TLSConfig, Timeout, Value,
 };
-use std::fs::File;
-use std::io::Write;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 use std::process::ExitCode;
 use std::time::Duration;
 use std::{env, io};
@@ -322,7 +320,7 @@ async fn main() -> ExitCode {
         return ExitCode::FAILURE;
     };
 
-    let profiling = Profiling::default();
+    let profiling = profile::Writer::default();
     let code = match &opts.mode {
         Action::Add(cmd) => run_add(&opts, cmd, &client).await,
         Action::Bench(cmd) => run_bench(&opts, cmd, client).await,
@@ -342,36 +340,6 @@ async fn main() -> ExitCode {
     }
 
     code
-}
-
-#[derive(Debug, Default)]
-pub struct Profiling {
-    profiler: Profiler,
-}
-
-impl Profiling {
-    pub fn finish<P>(&self, path: P)
-    where
-        P: AsRef<Path>,
-    {
-        if let Err(e) = self.profiler.proto().and_then(|b| Self::write_file(&path, &b)) {
-            tracing::warn!(message = "unable to collect and write pprof data", path = ?path.as_ref(), err = %e);
-        }
-    }
-
-    fn write_file<P>(path: P, contents: &[u8]) -> Result<(), MtopError>
-    where
-        P: AsRef<Path>,
-    {
-        if contents.is_empty() {
-            return Err(MtopError::configuration("mtop built without profiling support"));
-        }
-
-        let mut file = File::options().create(true).truncate(true).write(true).open(path)?;
-        file.write_all(contents)?;
-        file.flush()?;
-        Ok(file.sync_all()?)
-    }
 }
 
 async fn new_client(opts: &McConfig, servers: &[Server]) -> Result<MemcachedClient, MtopError> {
