+++ mtop/src/profile/mod.rs
@@ -1,9 +1,12 @@
 #[cfg(not(feature = "profile"))]
+mod nop;
+#[cfg(not(feature = "profile"))]
 pub use nop::Profiler;
+
+#[cfg(feature = "profile")]
+mod proto;
 #[cfg(feature = "profile")]
 pub use proto::Profiler;
 
-#[cfg(not(feature = "profile"))]
-mod nop;
-#[cfg(feature = "profile")]
-mod proto;
+mod writer;
+pub use writer::Writer;
